Extract inner classes to separate files
