+++ platform/base/core/src/test/java/com/peregrine/reference/impl/TreeTest.java
@@ -1,7 +1,5 @@
 package com.peregrine.reference.impl;
 
-import com.peregrine.reference.impl.ReferenceListerService.Node;
-import com.peregrine.reference.impl.ReferenceListerService.Tree;
 import org.junit.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
